Добавлен компонент на случай отсутствия предложений
